Remove header from reader view (#131)
* move ThemeProvider into ContextProvider

* use LayoutPage for all pages except reader

* add usage comments for LayoutPage and Page components

* update CHANGELOG
