+++ src/pages/book/[bookUrl].tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 import BookDetails from "components/bookDetails";
 import withAppProps, { AppProps } from "dataflow/withAppProps";
-import Page from "components/Page";
+import LayoutPage from "components/LayoutPage";
 import { NextPage, GetServerSideProps } from "next";
 
 const BookPage: NextPage<AppProps> = ({ library, error }) => {
   return (
-    <Page library={library} error={error}>
+    <LayoutPage library={library} error={error}>
       <BookDetails />
-    </Page>
+    </LayoutPage>
   );
 };
