+++ src/pages/loans.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 import { NextPage, GetServerSideProps } from "next";
-import Page from "components/Page";
+import LayoutPage from "components/LayoutPage";
 import withAppProps, { AppProps } from "dataflow/withAppProps";
 import MyBooks from "components/MyBooks";
 
 const MyBooksPage: NextPage<AppProps> = ({ library, error }) => {
   return (
-    <Page library={library} error={error}>
+    <LayoutPage library={library} error={error}>
       <MyBooks />
-    </Page>
+    </LayoutPage>
   );
 };
